Agent_creation/requirements.txt
Flask
python-dotenv
boto3
botocore
pandas
torch
transformers
openpyxl